Rename several items for clarity
- `script` is now `run`
- `run` is now `command`

This allows runnable shell commands to be consistent between variables
and tasks. Moving `run` to the outer layer also opens the possibility of
extending functionality in a way that aligns with plain English.

For example, pre-tasks can move into the `run` clause, and form a
pseudo-sentence such as "Run task: foo when exists: file". The same idea
works with `command`, ideally is a better abstraction.