Fix race condition for printing command output
The issue occurred because execCommand could return before all writes to
console occurred. Because these writes happen in a goroutine, it was
possible for the program to end with outputs still not written.

The -cover flag was added to the default test task to catch future
errors of this type.