Bump tool versions
The main changes here are the Go version bump and the mkdocs bump. The
rest of the changes are to support those.

With mkdocs in particular, we get a theme update for the built-in
readthedocs theme that lets us drop the rtd-dropdown package, and
updated code highlighting that I'm happier with than the custom pygments
highlighting that was being patched in before.