Fix issue executing logic during tab completions
This reprioritizes the logic in `main` such that safer behaviors are
prioritized over less safe behaviors. The most signficant is that no
behavior other than completion logic will run when the applcation is
triggered for the purpose of tab completions, but this also ensures that
passing the help or version flags take priority over all other
behaviors.