Ditch custom docker image for CI
The ability to run the tests locally as we would in CI was nice, but not
nice enough to warrant the additional overhead in having to build and
push new images manually before CI could be run. Instead, use docker
images that are suited to each stage.

Additionally bump Go, golangci-lint, and mkdocs versions. Hold off on
goreleaser for now.