Add support for dotenv files
While I'm not sure there is a formal dotenv specification, I've used a
library that's meant to be compatible with other language libraries,
which should make things pretty familiar and compatible.

I've also attempted to model the semantics used by Docker Compose for
file configuration, although 100% compatibility is probably not a
guarantee this project will ever make.

This is currently pending docs and release notes.