Migrate to mise for local/CI and bump versions
Normally I wouldn't want to introduce an extra dependency in CI, but I
think this would be extremely easy to undo if it ever started to cause
problems. The main reason to use it in actions is to avoid declaring
tool versions in two places that need to be kept in sync.

The major benefit here is that for tools like golangci-lint and
goreleaser, we now use an official release binary rather than building
from source, which was never officially supported or recommended. Not
that it's caused problems yet, but better to play it safe.

This also has the benefit of ensuring that the version of golangci-lint
that runs in my editor is the same as the version the project wants me
to use.