Refactor for better metadata in help text
As it was written before, task metadata was only available in help text
to a limited degree. This meant that propagating things like valid input
values, default values, and environment variables was not possible.
While this refactor does absolutely none of those things (and frankly,
complicates the code a bit), it makes them ultimately possible.