Rely more heavily on urfave/cli logic for flag prefixing
They don't expose everything I'd like them to directly, but we can do a
pretty good job of piecing it together.

Still pending is making sure that backticks work (or don't, but don't
break at the very least). But this refactor is better than what we had
before, I think.