Swap to doublestar for ** glob syntax
Even though technically `schemagen/*` worked for the use case I had,
what I actually wanted to specify was `schemagen/**/*`. The fact that
the directory was simple enough to not have nesting was incidental, and
not structural. Having to specify one pattern for each expected level of
indentation makes a lot less sense than bringing in a dependency.