Replace Directory/Dir with CfgPath
There are contexts in which we're going to need the full config path,
and not just the directory. Specifically, with regards to config-file
specific caching behavior, but really anything that uniquely identifies
the config we're using.

Since the directory can always be derived from the full config path,
let's just defer the indirection until we need it.