Always exit 1 on cache related errors
I was thinking there are some scenarios like failing to write the cache
after a task executes that maybe it would be fine to warn the user, but
not necessarily fail the task. But the test case consistency didn't tell
a great story about the first iteration, which got me thinking.

It's a lot easier to take something that used to be an error and let it
pass than it is to take something that used to be a warning and make it
fail. So no warnings for now. If I discover legitimate use cases for
allowing failures I can definitely revisit then.