Move Metadata into appcli
I realized the effective constructor for this type lived in a separate
package from the type itself. That seems like a reasonable enough signal
to refactor.

There are some exported things that can be unexported, but I figured it
was better to have a commit that was virtually just the move, and a
separate commit to clean up the things that can be cleaned up.