Improve "include" unmarshal error handling
There was a weird emergent multi-file error handling quirk that a test
did enforce, but that was definitely not what we wanted to do.

Rather than avoiding unwrapping errors to skip out on special casing one
error in particular, we want to intentionally opaque the unmarshal error
when it comes from an included file, such that it's not treated as one
candidate among many and instead forces the error with the most helpful
message to bubble up.